Added multiple Quentity of items on Cart flogic

+++ E-Commerce-Js/src/addToCard.js
@@ -1,34 +1,61 @@
 import { getCartProductFromLS } from "./getCartProducts";
 import { updateCartValue } from "./updateCartValue";
 
-export const addToCart=(event,id,stock)=>{
 
 
-let arrLocalStorageProduct=getCartProductFromLS();
+getCartProductFromLS();
+export const addToCart = (event, id, stock) => {
 
 
-    const currentProductElem=document.querySelector(`#card${id}`);
-    let quantity=currentProductElem.querySelector('.productQuantity').innerText;
-    let price=currentProductElem.querySelector('.productPrice').innerText;
+    let arrLocalStorageProduct = getCartProductFromLS();
+
+
+    const currentProductElem = document.querySelector(`#card${id}`);
+    let quantity = currentProductElem.querySelector('.productQuantity').innerText;
+    let price = currentProductElem.querySelector('.productPrice').innerText;
 
     // console.log(quantity,price);
 
-    price=price.replace ("₹","");
-    
-    price=Number(price*quantity);
-    quantity=Number(quantity);
+    price = price.replace("₹", "");
 
-    let UpdatedCart={id,quantity,price};
+
+    let existingProd = arrLocalStorageProduct.find((curProd) => {
+        return curProd.id === id;
+
+    });
+
+    if (existingProd && quantity > 1) {
+        quantity = Number(existingProd.quantity) + Number(quantity);
+        // alert("Item Already Exist");
+        price = Number(price * quantity);
+        let updatedCart = { id, quantity, price };
+        updatedCart = arrLocalStorageProduct.map((curProd) => {
+            return (curProd.id === id) ? updatedCart : curProd;
+        });
+        console.log(updatedCart)
+        localStorage.setItem('cartProductLS', JSON.stringify(updatedCart));
+
+
+    }
+
+    if (existingProd) {
+        return false;
+    }
+
+    price = Number(price * quantity);
+    quantity = Number(quantity);
+
+    let UpdatedCart = { id, quantity, price };
     arrLocalStorageProduct.push(UpdatedCart);
 
     //Or we can write it like 
     // arrLocalStorageProduct.push({id,quantity,price});
 
-    localStorage.setItem('cartProductLS',JSON.stringify(arrLocalStorageProduct));
+    localStorage.setItem('cartProductLS', JSON.stringify(arrLocalStorageProduct));
 
 
 
     //Update Cart Button Value
     updateCartValue(arrLocalStorageProduct);
-    
+
 }
